Extract runWithTimeout function to utils.kt
The runWithTimeout function initially present in CoverageCollector.kt has been refactored into a separate utils.kt file. With this change, the function is now more accessible and reused across different parts of the code. The function in CoverageCollector is replaced with a reference to the newly extracted runWithTimeoutOrNull function in the utils.kt file.